requirements.txt
asgiref==3.3.1
astroid==2.4.2
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.5
daphne==3.0.1
Django==3.1.6
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
ffmpeg-python==0.2.0
future==0.18.2
google-api-core==1.26.3
google-auth==1.28.0
google-cloud-speech==2.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
isort==5.7.0
lazy-object-proxy==1.4.3
libcst==0.3.17
Markdown==3.3.3
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.9
proto-plus==1.18.1
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
pylint==2.6.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pysensei @ file:///C:/Users/Anto/Desktop/virtualautomation/pysensei-2.0.7/pysensei-2.0.7-py3-none-any.whl
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted @ file:///C:/Users/Anto/Downloads/Twisted-20.3.0-cp39-cp39-win_amd64.whl
txaio==20.12.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
whitenoise==5.2.0
wrapt==1.12.1
zope.interface==5.2.0
